requirements.txt
two1
redis
netifaces
websocket-client
docker-py
flask